create base class for llm client
